+++ actix-server/examples/tcp-echo.rs
@@ -19,7 +19,7 @@ use std::{
 
 use actix_rt::net::TcpStream;
 use actix_server::Server;
-use actix_service::{fn_service, ServiceFactoryExt as _};
+use actix_service::{fn_factory, fn_service, ServiceExt as _};
 use bytes::BytesMut;
 use log::{error, info};
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
@@ -39,53 +39,62 @@ async fn main() -> io::Result<()> {
     // to return a service *factory*; so it can be created once per worker.
     Server::build()
         .bind("echo", addr, {
-            let count = Arc::clone(&count);
-            let num2 = Arc::clone(&count);
-
-            let svc = fn_service(move |mut stream: TcpStream| {
+            fn_factory::<_, (), _, _, _, _>(move || {
                 let count = Arc::clone(&count);
 
                 async move {
-                    let num = count.fetch_add(1, Ordering::SeqCst);
-                    let num = num + 1;
+                    let count = Arc::clone(&count);
+                    let count2 = Arc::clone(&count);
 
-                    let mut size = 0;
-                    let mut buf = BytesMut::new();
+                    let svc = fn_service(move |mut stream: TcpStream| {
+                        let count = Arc::clone(&count);
 
-                    loop {
-                        match stream.read_buf(&mut buf).await {
-                            // end of stream; bail from loop
-                            Ok(0) => break,
+                        let num = count.fetch_add(1, Ordering::SeqCst) + 1;
 
-                            // more bytes to process
-                            Ok(bytes_read) => {
-                                info!("[{}] read {} bytes", num, bytes_read);
-                                stream.write_all(&buf[size..]).await.unwrap();
-                                size += bytes_read;
+                        info!(
+                            "[{}] accepting connection from: {}",
+                            num,
+                            stream.peer_addr().unwrap()
+                        );
+
+                        async move {
+                            let mut size = 0;
+                            let mut buf = BytesMut::new();
+
+                            loop {
+                                match stream.read_buf(&mut buf).await {
+                                    // end of stream; bail from loop
+                                    Ok(0) => break,
+
+                                    // more bytes to process
+                                    Ok(bytes_read) => {
+                                        info!("[{}] read {} bytes", num, bytes_read);
+                                        stream.write_all(&buf[size..]).await.unwrap();
+                                        size += bytes_read;
+                                    }
+
+                                    // stream error; bail from loop with error
+                                    Err(err) => {
+                                        error!("Stream Error: {:?}", err);
+                                        return Err(());
+                                    }
+                                }
                             }
 
-                            // stream error; bail from loop with error
-                            Err(err) => {
-                                error!("Stream Error: {:?}", err);
-                                return Err(());
-                            }
+                            // send data down service pipeline
+                            Ok((buf.freeze(), size))
                         }
-                    }
+                    })
+                    .map_err(|err| error!("Service Error: {:?}", err))
+                    .and_then(move |(_, size)| {
+                        let num = count2.load(Ordering::SeqCst);
+                        info!("[{}] total bytes read: {}", num, size);
+                        async move { Ok(size) }
+                    });
 
-                    // send data down service pipeline
-                    Ok((buf.freeze(), size))
+                    Ok::<_, ()>(svc.clone())
                 }
             })
-            .map_err(|err| error!("Service Error: {:?}", err))
-            .and_then_send(move |(_, size)| {
-                let num = num2.load(Ordering::SeqCst);
-                info!("[{}] total bytes read: {}", num, size);
-                async move { Ok(size) }
-            });
-
-            let svc2 = svc.clone();
-
-            svc2
         })?
         .workers(2)
         .run()